volume work
git-svn-id: file:///home/lennart/svn/public/pulseaudio/trunk@42 fefdeb5f-60dc-0310-8127-8f9354f1896f

+++ src/sink.h
@@ -20,7 +20,7 @@ struct sink {
 
     struct source *monitor_source;
 
-    uint8_t volume;
+    uint32_t volume;
 
     void (*notify)(struct sink*sink);
     uint32_t (*get_latency)(struct sink *s);
